requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
google-genai>=0.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
PyJWT>=2.0.0
passlib[bcrypt]>=1.7.0
slowapi>=0.1.0
python-multipart>=0.0.6
